+++ sequdas_qc/Conf/config.ini
@@ -1,38 +1,38 @@
 [basic]
-write_logfile_details = 
-admin_email = 
-logfile_dir = 
-send_email =
-server_dir = 
+write_logfile_details = a
+admin_email = a
+logfile_dir = a
+send_email = a
+server_dir = a
 
 [reporter]
-reporter_ssh_host = 
-qc_dir = 
+reporter_ssh_host = a
+qc_dir = a
 
 [mysql_account]
-mysql_host = 
-mysql_user = 
-mysql_passwd = 
-mysql_db = 
+mysql_host = a
+mysql_user = a
+mysql_passwd = a
+mysql_db = a
 
 [email_account]
-gmail_user = 
-gmail_pass = 
+gmail_user = a 
+gmail_pass = a
 
 [kraken]
-db = 
+db = a
 
 [conda]
-krona = 
-interop = 
+krona = a
+interop = a
 
 [uploader]
-irida = 
+irida = a
 
 [genus]
-listeria = 
-e.coli = 
-escherichia = 
-shigella = 
-salmonella = 
-campylobacter
+listeria = 1
+e.coli = 1
+escherichia = 1
+shigella = 1
+salmonella = 1
+campylobacter = 1
