+++ sequdas_qc/cluster/fastqc.sh
