+++ sequdas_qc/cluster/kraken2.sh
